updated corrections

+++ controller.js
@@ -10,10 +10,11 @@ const { createNewUser,
     deleteParcel,
     getAllParcel,
     findUserParcel,
-    getUserId,
-    checkAdmin,
+    getUserStatusById,
+    checkSuperAdmin,
     checkIfUserExistById,
-    getAllParcelByUser
+    getAllParcelByUser,
+    getUserId
 } = require("./vergeService");
 
 const { validateEmail, validatePassword } = require("./token");
@@ -51,8 +52,10 @@ const signUpUser = async (req, res) => {
 
 const signUpAdmin = async (req, res) => {
     const { email } = req.body;
+    const id = res.locals.user.id;
     try {
         await checkIfUserDoesNotExistBefore(email);
+        await checkSuperAdmin(id);
         const result = await createNewAdmin(req.body);
         return res.status(201).json(result);
     } catch (e) {
@@ -120,14 +123,32 @@ const changeUserOrderStatus = async (req, res) => {
     }
 };
 
+const statusValidation = async (req, res, next) => {
+    const {status} = req.body;
+    if (status !== "processing" && status !== "shipped" && status !== "delivered"){
+        return res.status(401).json({
+            message: `You can not set status to '${status}'. Use 'processing', 'shipped' or 'delivered'.`
+        })
+    }
+    next();
+}
+
 const updateUserOrderDestination = async (req, res) => {
     const { id } = req.params;
     const user_id = res.locals.user.id;
+    const first_name = res.locals.user.first_name;
     try {
-        // await checkIfUserExistById(user_id);
-        await checkStatus(id);
+        const currentUserId = await getUserId(id);
+        if(currentUserId === user_id){
+            await checkStatus(id);
         const result = await updateOrderDestination(user_id, id, req.body);
-        return res.status(200).json(result)
+        return res.status(200).json(result);
+        }
+        if(currentUserId !== user_id){
+            return res.status(401).json({
+                message: `Dear ${first_name}, You can only update destination of an order created by you`,
+            });
+        }        
     } catch (e) {
         return res.status(e.code).json(e)
     }
@@ -156,9 +177,20 @@ const validateId = (req, res, next) => {
 const deleteUserParcel = async (req, res) => {
     const { id } = req.params;
     try {
-        await getUserId(id);
-        const result = await deleteParcel(id);
-        return res.status(200).json(result);
+        await getUserStatusById(id);
+        const currentUserId = await getUserId(id);
+        const user_id = res.locals.user.id;
+        const first_name = res.locals.user.first_name;
+        if(currentUserId === user_id){
+            const result = await deleteParcel(id);
+            return res.status(200).json(result);
+        }
+        if(currentUserId !== user_id){
+            return res.status(401).json({
+                message: `Dear ${first_name}, You can only delete an order created by you`,
+            });
+        }
+        
     } catch (e) {
         return res.status(e.code).json(e);
     }
@@ -208,5 +240,6 @@ module.exports = {
     deleteUserParcel,
     getAllUserParcelsByAdmin,
     getSpecificParcel,
-    getAllUsersParcel
+    getAllUsersParcel,
+    statusValidation
 }
